% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_affective.R
\docType{data}
\name{affective}
\alias{affective}
\title{Data from the Affective Style Questionnaire (ASQ - French Validation)}
\format{
A data frame with 1277 rows and 8 variables:
\describe{
  \item{Sex}{Sex (F or M)}
  \item{Birth_Season}{Season of birth}
  \item{Age}{Current age}
  \item{Salary}{Salary in euros}
  \item{Life_Satisfaction}{General life satisfaction}
  \item{Concealing}{Concealing score}
  \item{Adjusting}{Adjusting score}
  \item{Tolerating}{Tolerating score}
}
}
\usage{
affective
}
\description{
This is data from the French validation of the Affective Style Questionnaire.
}
\keyword{datasets}
