% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.anova.R
\name{analyze.anova}
\alias{analyze.anova}
\title{Analyze anova objects.}
\usage{
\method{analyze}{anova}(x, ...)
}
\arguments{
\item{x}{anova object.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Analyze anova objects.
}
\examples{
library(psycho)

df <- psycho::affective

x <- anova(lm(df$Tolerating ~ df$Salary))
x <- anova(lm(df$Tolerating ~ df$Salary * df$Sex))

summary(analyze(x))


}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
