% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.aov.R
\name{analyze.aov}
\alias{analyze.aov}
\title{Analyze aov objects.}
\usage{
\method{analyze}{aov}(x, ...)
}
\arguments{
\item{x}{aov object.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Analyze aov objects.
}
\examples{
library(psycho)

df <- psycho::affective

x <- aov(df$Tolerating ~ df$Sex)

summary(analyze(x))


}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
