% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.lmerMod.R
\name{analyze.lmerMod}
\alias{analyze.lmerMod}
\title{Analyze lmerMod objects.}
\usage{
\method{analyze}{lmerMod}(x, ...)
}
\arguments{
\item{x}{lmerMod object.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Analyze lmerMod objects.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
