% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.merMod.R
\name{analyze.merMod}
\alias{analyze.merMod}
\title{Analyze merMod objects.}
\usage{
\method{analyze}{merMod}(x, ...)
}
\arguments{
\item{x}{merMod object.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Analyze merMod objects.
}
\examples{
library(psycho)
require(lme4)
fit <- lme4::lmer(Sepal.Length ~ Sepal.Width + (1|Species), data=iris)

results <- analyze(fit)
summary(results)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
