% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{analyze}
\alias{analyze}
\title{Analyze objects.}
\usage{
analyze(x, ...)
}
\arguments{
\item{x}{object to analyze.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Analyze objects. See the documentation for your object's class:
\itemize{
 \item{\link[=analyze.stanreg]{analyze.stanreg}}
 \item{\link[=analyze.lmerModLmerTest]{analyze.merModLmerTest}}
 \item{\link[=analyze.glmerMod]{analyze.glmerMod}}
 \item{\link[=analyze.lm]{analyze.lm}}
 \item{\link[=analyze.glm]{analyze.glm}}
 }
 \itemize{
 \item{\link[=analyze.htest]{analyze.htest}}
 \item{\link[=analyze.aov]{analyze.aov}}
 }
\itemize{
 \item{\link[=analyze.fa]{analyze.fa}}
 \item{\link[=analyze.principal]{analyze.principal}}
 \item{\link[=analyze.lavaan]{analyze.lavaan}}
 \item{\link[=analyze.blavaan]{analyze.blavaan}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
