% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.stanreg.R
\name{analyze.stanreg}
\alias{analyze.stanreg}
\title{Analyze stanreg objects.}
\usage{
\method{analyze}{stanreg}(x, CI = 95, MEDP_step = 0.05, MEDP_min = 0.001,
  ...)
}
\arguments{
\item{x}{stanreg object.}

\item{CI}{Credible interval bounds.}

\item{MEDP_step}{The step used for computing Maximum Effect Direction Probability.}

\item{MEDP_min}{The lower bound of MEDP (the minimum effect).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Analyze stanreg objects.
}
\examples{
library(psycho)
require(rstanarm)
fit <- rstanarm::stan_glm(vs ~ mpg * cyl,
  data=mtcars,
  family = binomial(link = "logit"),
  prior=NULL)

 analyze(fit)

}
\author{
Dominique Makowski, \url{https://dominiquemakowski.github.io/}
}
