% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Compare a score to a parent population.}
\usage{
assess(patient, mean = 0, sd = 1, n = NULL, controls = NULL,
  verbose = TRUE, CI = 95, iter = 10000, color = "#2196F3",
  color_CI = "#E91E63", color_score = "black", color_size = 2)
}
\arguments{
\item{patient}{Single value (patient's score).}

\item{mean}{Mean of the control sample.}

\item{sd}{SD of the control sample.}

\item{n}{Size of the control sample.}

\item{controls}{Vector of values (control's scores).}

\item{verbose}{Print possible warnings.}

\item{CI}{Credible interval bounds.}

\item{iter}{Number of iterations.}

\item{color}{Main color of plot distribution.}

\item{color_CI}{Color of CI area.}

\item{color_score}{Color of the line representing the patient's score.}

\item{color_size}{Size of the line representing the patient's score.}
}
\value{
output
}
\description{
Compare a given score to a parent population.
}
\examples{
result <- assess(patient=124, mean=100, sd=15)
print(result)
plot(result)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
