% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.stanreg.R
\name{bayes_adj_R2}
\alias{bayes_adj_R2}
\title{Compute LOO-adjusted R2.}
\usage{
bayes_adj_R2(fit)
}
\arguments{
\item{fit}{A stanreg model.}
}
\description{
Compute LOO-adjusted R2.
}
\examples{
\dontrun{
library(psycho)
library(rstanarm)

data <- attitude
fit <- rstanarm::stan_glm(rating ~ advance + privileges, data=data)

bayes_adj_R2(fit)

}

}
\author{
\href{https://github.com/strengejacke}{Daniel Luedecke}
}
