% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_cor.R
\name{bayesian_cor}
\alias{bayesian_cor}
\title{Bayesian Correlation Matrix.}
\usage{
bayesian_cor(df)
}
\arguments{
\item{df}{The dataframe.}
}
\value{
A list of dataframes
}
\description{
Bayesian Correlation Matrix.
}
\examples{
library(psycho)
df <- psycho::affective
bayesian_cor(df)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
