% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_packages.R
\name{cite_packages}
\alias{cite_packages}
\title{Citations of loaded packages.}
\usage{
cite_packages(session)
}
\arguments{
\item{session}{A `devtools::sessionInfo()` object.}
}
\description{
Get the citations of loaded packages.
}
\examples{
\dontrun{
library(psycho)
cite_packages(sessionInfo())
}

}
\author{
\href{https://github.com/DominiqueMakowski}{Dominique Makowski}
}
