% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Multiple Correlations.}
\usage{
correlation(df, df2 = NULL, type = "full", method = "pearson",
  adjust = "holm")
}
\arguments{
\item{df}{The score.}

\item{df2}{The general population's mean.}

\item{type}{A character string indicating which correlation type is to be computed. One of "full" (default"), "partial" or "semi" for semi-partial correlations.}

\item{method}{A character string indicating which correlation coefficient is to be computed. One of "pearson" (default), "kendall", or "spearman" can be abbreviated.}

\item{adjust}{What adjustment for multiple tests should be used? ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). See \link[stats]{p.adjust} for details about why to use "holm" rather than "bonferroni").}
}
\value{
output
}
\description{
Compare a given score to a parent population.
}
\examples{
assess(124, mean=100, sd=15)

}
\author{
Dominique Makowski, \url{https://dominiquemakowski.github.io/}
}
