% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_random_effects.R
\name{extract_random_effects}
\alias{extract_random_effects}
\title{Find random effects in formula.}
\usage{
extract_random_effects(formula)
}
\arguments{
\item{formula}{Formula}
}
\description{
Find random effects in formula.
}
\examples{
library(psycho)
extract_random_effects("Y ~ X + (1|Group)")

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
