% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_combination.R
\name{find_best_combination}
\alias{find_best_combination}
\title{Returns the best combination of predictors.}
\usage{
find_best_combination(fit, ...)
}
\arguments{
\item{fit}{Model}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Returns the best combination of predictors.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
