% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_combination.stanreg.R
\name{find_best_combination.stanreg}
\alias{find_best_combination.stanreg}
\title{Returns the best combination of predictors based on LOO cross-validation indices.}
\usage{
\method{find_best_combination}{stanreg}(fit, interaction = TRUE, K = 10,
  k_treshold = NULL, ...)
}
\arguments{
\item{fit}{A stanreg object.}

\item{interaction}{Include interaction term.}

\item{K}{dupa}

\item{k_treshold}{dupa}

\item{...}{Arguments passed to or from other methods.}
}
\value{
list containing all combinations.
}
\description{
Returns the best combination of predictors based on LOO cross-validation indices.
}
\examples{
\dontrun{
library(psycho)
library(rstanarm)

data <- standardize(attitude)
fit <- rstanarm::stan_glm(rating ~ advance + privileges, data=data)

best <- find_best_combination(fit)
best_formula <- best$formula
best$table

# To deactivate Kfold evaluation
best <- find_best_combination(fit, K=0)
}

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
