% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_combinations.R
\name{find_combinations}
\alias{find_combinations}
\title{Generate all combinations.}
\usage{
find_combinations(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Generate all combinations.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
