% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{find_distance_cluster}
\alias{find_distance_cluster}
\title{Find the distance of a point with its kmean cluster.}
\usage{
find_distance_cluster(df, km)
}
\arguments{
\item{df}{Data}

\item{km}{kmean object.}
}
\description{
Find the distance of a point with its kmean cluster.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
