% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_highest_density_point.R
\name{find_highest_density_point}
\alias{find_highest_density_point}
\title{Find the Highest Density Point.}
\usage{
find_highest_density_point(x, precision = 1000)
}
\arguments{
\item{x}{Vector.}

\item{precision}{Number of points in density.}
}
\description{
Returns the Highest Density Point.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
