% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{find_matching_string}
\alias{find_matching_string}
\title{Fuzzy string matching.}
\usage{
find_matching_string(x, y, value = TRUE, step = 0.1, ignore.case = TRUE)
}
\arguments{
\item{x}{Strings.}

\item{y}{List of strings to be matched.}

\item{value}{Return value or the index of the closest string.}

\item{step}{Step by which decrease the distance.}

\item{ignore.case}{if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.}
}
\description{
Fuzzy string matching.
}
\examples{
library(psycho)
find_matching_string("Hwo rea ouy", c("How are you", "Not this word", "Nice to meet you"))
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
