% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{find_random_effects}
\alias{find_random_effects}
\title{Find random effects in formula.}
\usage{
find_random_effects(formula)
}
\arguments{
\item{formula}{Formula}
}
\description{
Find random effects in formula.
}
\examples{
library(psycho)
find_random_effects("Y ~ X + (1|Group)")
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
