% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_season.R
\name{find_season}
\alias{find_season}
\title{Find season of dates.}
\usage{
find_season(date)
}
\arguments{
\item{date}{Array of dates. Must cover the 4 seasons.}
}
\value{
season
}
\description{
Returns the season of an array of dates.
}
\examples{
dates <- c("2017-02-15", "2017-05-15", "2017-08-15", "2017-11-15")
find_season(dates)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
