% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{format_bf}
\alias{format_bf}
\title{Bayes factor formatting}
\usage{
format_bf(bf, max = 100)
}
\arguments{
\item{bf}{Bayes Factor.}

\item{max}{Treshold for maximum.}
}
\description{
Bayes factor formatting
}
