% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_digit.R
\name{format_digit}
\alias{format_digit}
\title{Formatting}
\usage{
format_digit(x, digits = 2)
}
\arguments{
\item{x}{number.}

\item{digits}{number of significant digits.}
}
\description{
Formatting
}
\examples{

format_digit(1.20)
format_digit(1.2)
format_digit(1.2012313)
format_digit(0.0045)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
