% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{format_formula}
\alias{format_formula}
\title{Clean and format formula.}
\usage{
format_formula(formula)
}
\arguments{
\item{formula}{formula}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Clean and format formula.
}
\examples{
library(psycho)
library(lme4)

fit <- lme4::glmer(vs ~ wt + (1 | gear), data = mtcars, family = "binomial")
fit <- lm(hp ~ wt, data = mtcars)

format_formula(get_formula(fit))
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
