% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{format_loadings}
\alias{format_loadings}
\title{Format the loadings of a factor analysis.}
\usage{
format_loadings(x, labels = NULL)
}
\arguments{
\item{x}{An psych object.}

\item{labels}{Supply a additional column with e.g. item labels.}
}
\description{
Format the loadings of a factor analysis.
}
\examples{
\dontrun{
library(psycho)

x <- psych::fa(psych::Thurstone.33, 2)
format_loadings(x)
}

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
