% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_p}
\alias{format_p}
\title{Format p values.}
\usage{
format_p(pvalues, stars = TRUE)
}
\arguments{
\item{pvalues}{P values (scalar or vector).}

\item{stars}{Add stars.}
}
\description{
Format p values.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
