% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrasts.lmerMod.R
\name{get_contrasts.lmerMod}
\alias{get_contrasts.lmerMod}
\title{Compute estimated marginal means and contrasts from lmerMod models.}
\usage{
\method{get_contrasts}{lmerMod}(fit, formula, adjust = "tukey", ...)
}
\arguments{
\item{fit}{A lmerMod model.}

\item{formula}{A character vector (formula like format, i.e., including
interactions or nesting terms) specifying the names of the predictors over which EMMs are desired.}

\item{adjust}{P value adjustment method. Default is "tukey". Can be "holm",
"hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{...}{Arguments passed to or from other methods.}
}
\value{
list with estimated marginal means (95% CI) and contrasts.
}
\description{
Compute estimated marginal means and contrasts from a lmerMod models.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
