% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrasts.merMod.R
\name{get_contrasts.merMod}
\alias{get_contrasts.merMod}
\title{Compute estimated marginal means and contrasts from merMod models.}
\usage{
\method{get_contrasts}{merMod}(fit, formula, adjust = "tukey", ...)
}
\arguments{
\item{fit}{A merMod model.}

\item{formula}{A character vector (formula like format, i.e., including
interactions or nesting terms) specifying the names of the predictors over which EMMs are desired.}

\item{adjust}{P value adjustment method. Default is "tukey". Can be "holm",
"hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{...}{Arguments passed to or from other methods.}
}
\value{
list with estimated marginal means (95% CI) and contrasts.
}
\description{
Compute estimated marginal means and contrasts from a merMod models.
}
\examples{
\dontrun{
library(psycho)
require(lme4)
fit <- lme4::lmer(Adjusting ~ Birth_Season + (1|Salary), data=affective)

contrasts <- get_contrasts(fit, formula="Birth_Season", adjust="tukey")
contrasts$means
contrasts$contrasts
}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
