% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrasts.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Get Marginal Means and Contrasts.}
\usage{
get_contrasts(fit, ...)
}
\arguments{
\item{fit}{Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute estimated marginal means and contrasts from models. See the
documentation for your model's class:
\itemize{
 \item{\link[=get_contrasts.stanreg]{get_contrasts.stanreg}}
 \item{\link[=get_contrasts.lmerModLmerTest]{get_contrasts.lmerModLmerTest}}
 \item{\link[=get_contrasts.glmerMod]{get_contrasts.glmerMod}}
 \item{\link[=get_contrasts.lm]{get_contrasts.lm}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
