% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrasts.stanreg.R
\name{get_contrasts.stanreg}
\alias{get_contrasts.stanreg}
\title{Compute estimated marginal means and contrasts from stanreg models.}
\usage{
\method{get_contrasts}{stanreg}(fit, specs, prob = 0.9, ...)
}
\arguments{
\item{fit}{A stanreg model.}

\item{specs}{A character vector specifying the names of the predictors over which EMMs are desired. specs may also be a formula or a list (optionally named) of valid specs. Use of formulas is described in the Details section below. See ?emmeans::emmeans.}

\item{prob}{A numeric scalar in the interval (0,1) giving the target probability content of the intervals. The nominal probability content of the intervals is the multiple of 1/nrow(obj) nearest to prob.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
list with estimated marginal means and contrasts.
}
\description{
Compute estimated marginal means and contrasts from a stanreg models.
}
\examples{
\dontrun{
library(psycho)
require(rstanarm)
fit <- rstanarm::stan_glm(vs ~ mpg * cyl, data=mtcars)

predicted <- get_predicted(fit)
}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
