% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_graph.fa}
\alias{get_graph.fa}
\title{Get graph data from factor analysis.}
\usage{
\method{get_graph}{fa}(fit, threshold_Coef = NULL, digits = 2, ...)
}
\arguments{
\item{fit}{psych::fa object.}

\item{threshold_Coef}{Omit all links with a Coefs below this value.}

\item{digits}{Edges' labels rounding.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A list containing nodes and edges data to be used by `tidygraph::tbl_graph()`.
}
\description{
Get graph data from fa objects.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
