% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph.R
\name{get_graph}
\alias{get_graph}
\title{Get graph data.}
\usage{
get_graph(fit, ...)
}
\arguments{
\item{fit}{Object from which to extract the graph data.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
To be used with tidygraph::tbl_graph. See the documentation for your object's class:
\itemize{
 \item{\link[=get_graph.lavaan]{get_graph.lavaan}}
 \item{\link[=get_graph.fa]{get_graph.fa}}
 \item{\link[=get_graph.psychobject_correlation]{get_graph.psychobject_correlation}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
