% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_info}
\alias{get_info}
\title{Get information about objects.}
\usage{
get_info(x, ...)
}
\arguments{
\item{x}{object.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Get information about models.
}
\examples{
library(psycho)
library(lme4)

fit <- lme4::glmer(vs ~ wt + (1 | gear), data = mtcars, family = "binomial")

info <- get_info(fit)
info
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
