% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_loadings_max}
\alias{get_loadings_max}
\title{Get loadings max.}
\usage{
get_loadings_max(loadings)
}
\arguments{
\item{loadings}{Formatted loadings.}
}
\description{
Get loadings max.
}
\examples{
\dontrun{
library(psycho)

x <- psych::fa(psych::Thurstone.33, 2)
get_loadings_max(format_loadings(x)$loadings)
}

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
