% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predicted.R
\name{get_predicted}
\alias{get_predicted}
\title{Compute predicted values from models.}
\usage{
get_predicted(fit, ...)
}
\arguments{
\item{fit}{Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute predicted values from models. See the
documentation for your model's class:
\itemize{
 \item{\link[=get_predicted.stanreg]{get_predicted.stanreg}}
 \item{\link[=get_predicted.merMod]{get_predicted.merMod}}
 \item{\link[=get_predicted.lm]{get_predicted.lm}}
 \item{\link[=get_predicted.glm]{get_predicted.glm}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
