% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_R2.R
\name{get_R2.lm}
\alias{get_R2.lm}
\title{R2 and adjusted R2 for Linear Models.}
\usage{
\method{get_R2}{lm}(fit, ...)
}
\arguments{
\item{fit}{A linear model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
R2 and adjusted R2 for Linear Models.
}
\examples{
\dontrun{
library(psycho)

fit <- lm(Tolerating ~ Adjusting, data = psycho::affective)

get_R2(fit)
}

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
