% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_R2}
\alias{get_R2}
\title{Get Indices of Explanatory Power.}
\usage{
get_R2(fit, ...)
}
\arguments{
\item{fit}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
See the documentation for your object's class:
\itemize{
\item{\link[=get_R2.lm]{get_R2.lm}}
\item{\link[=get_R2.glm]{get_R2.glm}}
\item{\link[=get_R2.stanreg]{get_R2.stanreg}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
