% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden.R
\name{golden}
\alias{golden}
\title{Golden Ratio.}
\usage{
golden(x = 1)
}
\arguments{
\item{x}{A number to be multiplied by the golden ratio. The default (x=1) returns the value of the golden ratio.}
}
\description{
Returns the golden ratio (1.618034...).
}
\examples{
library(psycho)

golden()
golden(8)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
