% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{HDImax}
\alias{HDImax}
\title{Highest Density Intervals (HDI)}
\usage{
HDImax(x, prob = 0.95)
}
\arguments{
\item{x}{A vector of values from a probability distribution (e.g., posterior probabilities from MCMC sampling).}

\item{prob}{Scalar between 0 and 1, indicating the mass within the credible interval that is to be estimated.}
}
\description{
See \link[=HDI]{HDI}
}
