% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Illustration of Hello world}
\usage{
hello(method2order, cex, mar)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Change something
prints hellow world \code{brocolors}
Use `devtools::document()`
devtools::build_vignettes()
}
\details{
library(devtools)
use_testthat()
use_test("your_function")
}
\examples{
hello()

}
