% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{interpret_blavaan}
\alias{interpret_blavaan}
\title{Interpret fit measures of blavaan objects}
\usage{
interpret_blavaan(fit, indices = c("BIC", "DIC", "WAIC", "LOOIC"), ...)
}
\arguments{
\item{fit}{A blavaan model.}

\item{indices}{Vector of strings indicating which indices to report. Only works for bayesian objects for now.}

\item{...}{Other arguments.}
}
\description{
Interpret fit measures of blavaan objects
}
