% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{interpret_d}
\alias{interpret_d}
\title{Standardized difference (Cohen's d) interpreation.}
\usage{
interpret_d(x, direction = FALSE, rules = "cohen1988")
}
\arguments{
\item{x}{Standardized difference.}

\item{direction}{Return direction.}

\item{rules}{Can be "cohen1988" (default), "sawilowsky2009", or a custom list.}
}
\description{
Interpret d with a set of rules.
}
\examples{
library(psycho)
interpret_d(-0.42)
interpret_d(-0.62)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
