% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_lavaan.R
\name{interpret_lavaan}
\alias{interpret_lavaan}
\title{Interpret fit measures of lavaan or blavaan objects}
\usage{
interpret_lavaan(fit, ...)
}
\arguments{
\item{fit}{lavaan or blavaan object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Interpret fit measures of lavaan or blavaan objects
}
