% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{interpret_omega_sq}
\alias{interpret_omega_sq}
\title{Omega Squared Interpretation}
\usage{
interpret_omega_sq(x, rules = "field2013")
}
\arguments{
\item{x}{Omega Squared.}

\item{rules}{Can be "field2013" (default), or a custom list.}
}
\description{
Return the interpretation of Omegas Squared.
}
\examples{
library(psycho)
interpret_omega_sq(x = 0.05)
}
\references{
\itemize{
 \item{Field, A (2013) Discovering statistics using IBM SPSS Statistics. Fourth Edition. Sage:London.}
 }
}
\seealso{
http://imaging.mrc-cbu.cam.ac.uk/statswiki/FAQ/effectSize
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
