% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_r.R
\name{interpret_r}
\alias{interpret_r}
\title{Return the interpretation of a correlation coefficient following Evans (1996).}
\usage{
interpret_r(x, direction = TRUE, strength = TRUE)
}
\arguments{
\item{x}{Correlation coefficient.}

\item{direction}{Return direction.}

\item{strength}{Return strength.}
}
\value{
The interpretation following Evans (1996).
}
\description{
Return the interpretation of a correlation coefficient following Evans (1996).
}
\examples{
library(psycho)
interpret_r(-0.42)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
