% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_R2.R
\name{interpret_R2}
\alias{interpret_R2}
\title{R2 interpreation.}
\usage{
interpret_R2(x, rules = "cohen1988")
}
\arguments{
\item{x}{Value.}

\item{rules}{Can be "cohen1988" (default), "chin1998" or "hair2013", or a custom list.}
}
\description{
Interpret R2 with a set of rules.
}
\examples{
library(psycho)
interpret_R2(x=0.42)
interpret_R2(x=c(0.42, 0.2, 0.9, 0))

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
