% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_r_posterior.R
\name{interpret_r_posterior}
\alias{interpret_r_posterior}
\title{Return the interpretation of a posterior distribution of correlation coefficient.}
\usage{
interpret_r_posterior(posterior)
}
\arguments{
\item{posterior}{Posterior distribution of correlation coefficient.}
}
\value{
The interpretation.
}
\description{
Return the interpretation of a posterior distribution of correlation coefficient.
}
\examples{
library(psycho)
posterior <- rnorm(1000, 0.5, 0.5)
interpret_r_posterior(posterior)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
