% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_RMSEA.R
\name{interpret_RMSEA}
\alias{interpret_RMSEA}
\title{RMSEA interpreation.}
\usage{
interpret_RMSEA(x, rules = "awang2012")
}
\arguments{
\item{x}{RMSEA.}

\item{rules}{Can be "awang2012", or a custom list.}
}
\description{
Interpret RMSEA with a set of rules.
}
\examples{
library(psycho)
interpret_RMSEA(0.04)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
