% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.mixed.R
\name{is.mixed}
\alias{is.mixed}
\title{Check if model includes random effects.}
\usage{
is.mixed(fit, ...)
}
\arguments{
\item{fit}{Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Check if model is mixed. See the
documentation for your model's class:
\itemize{
 \item{\link[=is.mixed.stanreg]{is.mixed.stanreg}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
