% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{is.mixed.stanreg}
\alias{is.mixed.stanreg}
\title{Check if model includes random effects.}
\usage{
\method{is.mixed}{stanreg}(fit, ...)
}
\arguments{
\item{fit}{Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Check if model is mixed.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
