% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychobject.R
\name{is.psychobject}
\alias{is.psychobject}
\title{Creates or tests for objects of mode "psychobject".}
\usage{
is.psychobject(x)
}
\arguments{
\item{x}{an arbitrary R object.}
}
\description{
Creates or tests for objects of mode "psychobject".
}
