% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{odds_to_d}
\alias{odds_to_d}
\title{(Log) odds ratio to Cohen's d}
\usage{
odds_to_d(x, log = TRUE)
}
\arguments{
\item{x}{Odds ratio.}

\item{log}{Are these log odds ratio?}
}
\description{
(Log) odds ratio to Cohen's d.
}
\examples{
library(psycho)
odds_to_d(x = 2)
}
\references{
\itemize{
 \item{Sánchez-Meca, J., Marín-Martínez, F., & Chacón-Moscoso, S. (2003). Effect-size indices for dichotomized outcomes in meta-analysis. Psychological methods, 8(4), 448.}
 }
}
\seealso{
https://www.meta-analysis.com/downloads/Meta-analysis%20Converting%20among%20effect%20sizes.pdf
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
