% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_to_probs.R
\name{odds_to_probs}
\alias{odds_to_probs}
\title{Convert (log)odds to probabilies.}
\usage{
odds_to_probs(odds, log = TRUE)
}
\arguments{
\item{odds}{Odds value(s).}

\item{log}{Are these Log odds (such as in logistic models)?}
}
\description{
Convert (log)odds to probabilies.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
