% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.aov.R
\name{omega_sq}
\alias{omega_sq}
\title{Partial Omega Squared.}
\usage{
omega_sq(x, partial = TRUE)
}
\arguments{
\item{x}{aov object.}

\item{partial}{Return partial omega squared.}
}
\value{
output
}
\description{
Partial Omega Squared.
}
\examples{
library(psycho)

df <- psycho::affective

x <- aov(df$Tolerating ~ df$Salary)
x <- aov(df$Tolerating ~ df$Salary * df$Sex)

omega_sq(x)

}
\seealso{
http://stats.stackexchange.com/a/126520
}
\author{
Arnoud Plantinga
}
