% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{percentile}
\alias{percentile}
\title{Transform z score to percentile.}
\usage{
percentile(z_score)
}
\arguments{
\item{z_score}{Z score.}
}
\description{
Transform z score to percentile.
}
\examples{
library(psycho)
percentile(-1.96)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
