% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{percentile_to_z}
\alias{percentile_to_z}
\title{Transform a percentile to a z score.}
\usage{
percentile_to_z(percentile)
}
\arguments{
\item{percentile}{Percentile}
}
\description{
Transform a percentile to a z score.
}
\examples{
library(psycho)
percentile_to_z(95)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
