% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.psycho.R
\name{plot.psycho}
\alias{plot.psycho}
\title{Plot the results.}
\usage{
\method{plot}{psycho}(x, ...)
}
\arguments{
\item{x}{A psycho class object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Plot the results.
}
\author{
Dominique Makowski, \url{https://dominiquemakowski.github.io/}
}
