% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.psychobject.R
\name{plot.psychobject}
\alias{plot.psychobject}
\title{Plot the results.}
\usage{
\method{plot}{psychobject}(x, ...)
}
\arguments{
\item{x}{A psychobject class object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Plot the results.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
