% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.lavaan.R
\name{plot_lavaan}
\alias{plot_lavaan}
\title{Plot lavaan (SEM or CFA) objects.}
\usage{
plot_lavaan(fit, links = c("Regression"), threshold_p = NULL,
  threshold_Coef = NULL)
}
\arguments{
\item{fit}{lavaan object.}

\item{links}{Which links to include? A list including at least one of "Regression", "Loading" or "Correlation".}

\item{threshold_p}{Omit all links with a p value below this value.}

\item{threshold_Coef}{Omit all links with a Coefs below this value.}
}
\value{
A list containing nodes and edges data to be used by `tidygraph::tbl_graph()`.
}
\description{
Plot lavaan (SEM or CFA) objects.
}
\seealso{
https://www.researchgate.net/post/Whats_the_standard_of_fit_indices_in_SEM
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
