% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot loadings.}
\usage{
plot_loadings(loadings)
}
\arguments{
\item{loadings}{Loadings by variable.}
}
\description{
Plot loadings.
}
\examples{
\dontrun{
library(psycho)

x <- psych::fa(psych::Thurstone.33, 2)
plot_loadings(format_loadings(x)$loadings)
}

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
