% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.psycho.R
\name{print.psycho}
\alias{print.psycho}
\title{Print the results.}
\usage{
\method{print}{psycho}(x, ...)
}
\arguments{
\item{x}{A psycho class object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print the results.
}
