% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_R2.R
\name{R2_LOO_Adjusted}
\alias{R2_LOO_Adjusted}
\title{Compute LOO-adjusted R2.}
\usage{
R2_LOO_Adjusted(fit)
}
\arguments{
\item{fit}{A stanreg model.}
}
\description{
Compute LOO-adjusted R2.
}
\examples{
\dontrun{
library(psycho)
library(rstanarm)

data <- attitude
fit <- rstanarm::stan_glm(rating ~ advance + privileges, data=data)

R2_LOO_Adjusted(fit)

}

}
\author{
\href{https://github.com/strengejacke}{Daniel Luedecke}
}
