% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\name{refdata}
\alias{refdata}
\title{Create a reference grid.}
\usage{
refdata(df, target = "all", length.out = 10, type = "reference",
  fixed = "mean")
}
\arguments{
\item{df}{The dataframe.}

\item{target}{String or list of strings to indicate target columns. Can be "all".}

\item{length.out}{Length of numeric target variables.}

\item{type}{"combinations" or "reference". Type of summary for factors.}

\item{fixed}{String indicating the type of summary for numerics.}
}
\description{
Create a reference grid.
}
\examples{
library(psycho)

df <- psycho::affective
newdata <- refdata(df, target="Sex")
newdata <- refdata(df, target="Sex", type="combinations")
newdata <- refdata(df, target=c("Sex", "Salary", "Tolerating"), length.out=3)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
