% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_cols.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Remove empty columns..}
\usage{
remove_empty_cols(df)
}
\arguments{
\item{df}{Dataframe.}
}
\description{
Removes all columns containing ony NaNs.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
