% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers.}
\usage{
remove_outliers(df, target, threshold = qnorm(0.95),
  direction = "both")
}
\arguments{
\item{df}{Dataframe.}

\item{target}{String or list of strings of variables}

\item{threshold}{The z-score value (deviation of SD) by which to consider outliers.}

\item{direction}{Can be "both", "upper" or "lower".}
}
\description{
Removes outliers (with the z-score method only for now).
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
