% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_file.R
\name{save_file}
\alias{save_file}
\title{Save the results to a file.}
\usage{
save_file(x, file = "results")
}
\arguments{
\item{x}{A psycho class object.}

\item{file}{The Filename.}
}
\description{
Save the results to a file.
}
\author{
Dominique Makowski, \url{https://dominiquemakowski.github.io/}
}
