% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.numeric}
\alias{standardize.numeric}
\title{Standardize (scale and reduce) numeric variables.}
\usage{
\method{standardize}{numeric}(x, normalize = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{normalize}{Will perform a normalization instead of a standardization. This scales all numeric variables in the range 0 - 1.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Standardize (Z-score, "normalize") a vector.
}
\examples{
standardize(x = c(1, 4, 6, 2))
standardize(x = c(1, 4, 6, 2), normalize = TRUE)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
