% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{standardize}
\alias{standardize}
\title{Standardize.}
\usage{
standardize(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Standardize objects. See the documentation for your object's class:
\itemize{
\item{\link[=standardize.numeric]{standardize.numeric}}
\item{\link[=standardize.data.frame]{standardize.data.frame}}
\item{\link[=standardize.stanreg]{standardize.stanreg}}
\item{\link[=standardize.lm]{standardize.lm}}
\item{\link[=standardize.glm]{standardize.glm}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
