% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.psycho.R
\name{summary.psycho}
\alias{summary.psycho}
\title{Print the results.}
\usage{
\method{summary}{psycho}(object, ...)
}
\arguments{
\item{object}{A psycho class object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print the results.
}
\author{
Dominique Makowski, \url{https://dominiquemakowski.github.io/}
}
