% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychobject.R
\name{summary.psychobject}
\alias{summary.psychobject}
\title{Print the results.}
\usage{
\method{summary}{psychobject}(object, round = NULL, ...)
}
\arguments{
\item{object}{A psychobject class object.}

\item{round}{Round the ouput.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print the results.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
