% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{values}
\alias{values}
\title{Extract values as list.}
\usage{
values(x)
}
\arguments{
\item{x}{A psychobject class object.}
}
\description{
Extract values as list.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
