% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.glmerMod.R
\name{analyze.glmerMod}
\alias{analyze.glmerMod}
\title{Analyze glmerMod objects.}
\usage{
\method{analyze}{glmerMod}(x, ...)
}
\arguments{
\item{x}{merModLmerTest object.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Analyze glmerMod objects.
}
\examples{
library(psycho)
library(lme4)

fit <- lme4::glmer(vs ~ wt + (1|gear), data=mtcars, family="binomial")

results <- analyze(fit)
summary(results)
print(results)



}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
