% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{analyze.lm}
\alias{analyze.lm}
\title{Analyze lm objects.}
\usage{
\method{analyze}{lm}(x, CI = 95, effsize_rules = "cohen1988", ...)
}
\arguments{
\item{x}{lm object.}

\item{CI}{Confidence interval bounds. Set to NULL turn off their computation.}

\item{effsize_rules}{Grid for effect size interpretation. See \link[=interpret_d]{interpret_d}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Analyze lm objects.
}
\examples{
library(psycho)
fit <- lm(Sepal.Length ~ Sepal.Width, data = iris)
fit <- lm(Sepal.Length ~ Sepal.Width * Species, data = iris)

results <- analyze(fit)
summary(results)
print(results)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
