% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.merModLmerTest.R
\name{analyze.merModLmerTest}
\alias{analyze.merModLmerTest}
\title{Analyze merModLmerTest objects.}
\usage{
\method{analyze}{merModLmerTest}(x, ...)
}
\arguments{
\item{x}{merModLmerTest object.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Analyze merModLmerTest objects.
}
\examples{
library(psycho)

}
\author{
Dominique Makowski, \url{https://dominiquemakowski.github.io/}
}
