% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze objects.}
\usage{
analyze(x, ...)
}
\arguments{
\item{x}{object to analyze.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Analyze objects. See the documentation for your object's class:
\itemize{
 \item{\link[=analyze.stanreg]{analyze.stanreg}}
 \item{\link[=analyze.lmerModLmerTest]{analyze.merModLmerTest}}
 \item{\link[=analyze.glmerMod]{analyze.glmerMod}}
 \item{\link[=analyze.lm]{analyze.lm}}
 }
\itemize{
 \item{\link[=analyze.fa]{analyze.fa}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
