% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.stanreg.R
\name{analyze.stanreg}
\alias{analyze.stanreg}
\title{Analyze stanreg objects.}
\usage{
\method{analyze}{stanreg}(x, CI = 90, effsize = FALSE,
  bayes_factor = FALSE, overlap = FALSE, ...)
}
\arguments{
\item{x}{A stanreg model.}

\item{CI}{Credible interval bounds.}

\item{effsize}{Compute Effect Sizes according to Cohen (1988)? Your outcome variable must be standardized.}

\item{bayes_factor}{Compute a Bayesian t-test to quantify the probability that each effect is different from 0.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
output
}
\description{
Analyze stanreg objects.
}
\examples{
\dontrun{
library(psycho)
library(rstanarm)

data <- standardize(attitude)
fit <- rstanarm::stan_glm(rating ~ advance + privileges, data=data)

results <- analyze(fit, effsize=TRUE, bayes_factor=TRUE)
summary(results)
plot(results)
print(results)

fit <- rstanarm::stan_glmer(Sepal.Length ~ Sepal.Width + (1|Species), data=iris)
results <- analyze(fit)
summary(results)
}

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
