% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Compare a score to a parent population.}
\usage{
assess(score, mean = 0, sd = 1, linecolor = "#E91E63",
  fillcolor = "#2196F3", verbose = T)
}
\arguments{
\item{score}{The score.}

\item{mean}{The general population's mean.}

\item{sd}{The general population's standart deviation.}

\item{linecolor}{The colour of the vertical line.}

\item{fillcolor}{The colour of the density plot.}

\item{verbose}{Print possible warnings.}
}
\value{
output
}
\description{
Compare a given score to a parent population.
}
\examples{
rez <- assess(124, mean=100, sd=15)

}
\author{
Dominique Makowski, \url{https://dominiquemakowski.github.io/}
}
