% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawford.test.R
\name{crawford.test}
\alias{crawford.test}
\title{Crawford-Garthwaite (2007) Bayesian test for single-case analysis.}
\usage{
crawford.test(patient, controls = NULL, mean = NULL, sd = NULL,
  n = NULL, CI = 95, iter = 10000, color = "#2196F3",
  color_CI = "#E91E63", color_score = "black", color_size = 2)
}
\arguments{
\item{patient}{Single value (patient's score).}

\item{controls}{Vector of values (control's scores).}

\item{mean}{Mean of the control sample.}

\item{sd}{SD of the control sample.}

\item{n}{Size of the control sample.}

\item{CI}{Credible interval bounds.}

\item{iter}{Number of iterations.}

\item{color}{Main color of plot distribution.}

\item{color_CI}{Color of CI area.}

\item{color_score}{Color of the line representing the patient's score.}

\item{color_size}{Size of the line representing the patient's score.}
}
\description{
Neuropsychologists often need to compare a single case to a small control group. However, the standard two-sample t-test does not work because the case is only one observation. Crawford and Garthwaite (2012) demonstrate that the Crawford-Garthwaite (2007) Bayesian test is a better approach than other commonly-used alternatives.
.
}
\details{
The p value obtained when this test is used to test significance also simultaneously provides a point estimate of the abnormality of the patient’s score; for example if the one-tailed probability is .013 then we know that the patient’s score is significantly (p , .05) below the control mean and that it is estimated that 1.3% of the control population would obtain a score lower than the patient’s. As for the credible interval interpretation, we could say that there is a 95% probability that the true level of abnormality of the patient’s score lies within the stated limits, or that There is 95% confidence that the percentage of people who have a score lower than the patient’s is between 0.01% and 6.66%.
}
\examples{
library(psycho)

crawford.test(patient = 10, controls = c(0, -2, 5, 2, 1, 3, -4, -2))
test <- crawford.test(patient = 7, controls = c(0, -2, 5, -6, 0, 3, -4, -2))
plot(test)

}
\author{
Dominique Makowski
}
