% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{find_best_model}
\alias{find_best_model}
\title{Returns the best model.}
\usage{
find_best_model(fit, ...)
}
\arguments{
\item{fit}{Model}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Returns the best model. See the
documentation for your model's class:
\itemize{
 \item{\link[=find_best_model.stanreg]{find_best_model.stanreg}}
 \item{\link[=find_best_model.lmerModLmerTest]{find_best_model.lmerModLmerTest}}
 }
}
\seealso{
\code{\link{find_best_model.stanreg}}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
