% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_digit}
\alias{format_digit}
\title{Format digits.}
\usage{
format_digit(x, digits = 2, null_treshold = 0.001)
}
\arguments{
\item{x}{A digit.}

\item{digits}{Number of significant digits.}

\item{null_treshold}{Treshold below which return 0.}
}
\description{
Format digits.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
