% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{format_p}
\alias{format_p}
\title{Format p values.}
\usage{
format_p(pvalues, stars = TRUE, stars_only = FALSE)
}
\arguments{
\item{pvalues}{p values (scalar or vector).}

\item{stars}{Add significance stars.}

\item{stars_only}{Return only significance stars.}
}
\description{
Format p values.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
