% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{Format p values.}
\usage{
format_p(pvalues)
}
\arguments{
\item{pvalues}{P values (scalar or vector).}
}
\description{
Format p values.
}
\author{
Dominique Makowski, \url{https://dominiquemakowski.github.io/}
}
