% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_string}
\alias{format_string}
\title{Tidyverse-friendly sprintf.}
\usage{
format_string(x, fmt, ...)
}
\arguments{
\item{x}{Values.}

\item{fmt}{A character vector of format strings, each of up to 8192 bytes.}
}
\description{
Tidyverse-friendly sprintf.
}
