% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.fa.R
\name{get_cfa_model}
\alias{get_cfa_model}
\title{Get CFA model.}
\usage{
get_cfa_model(loadings_max)
}
\arguments{
\item{loadings_max}{Max loadings by variable.}
}
\description{
Get CFA model.
}
\examples{
\dontrun{
library(psycho)

x <- psych::fa(psych::Thurstone.33, 2)
get_cfa_model(format_loadings(x)$max)
}


}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
