% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrasts.R
\name{get_contrasts.lmerMod}
\alias{get_contrasts.lmerMod}
\title{Compute estimated contrasts from models.}
\usage{
\method{get_contrasts}{lmerMod}(fit, formula = NULL, CI = 95,
  adjust = "tukey", ...)
}
\arguments{
\item{fit}{A frequentist model.}

\item{formula}{A character vector (formula like format, i.e., including
interactions or nesting terms) specifying the names of the predictors over which EMMs are desired.}

\item{CI}{Determine the confidence or credible interval bounds.}

\item{adjust}{P value adjustment method for frequentist models. Default is "tukey". Can be "holm",
"hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none".}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute estimated contrasts from models.
}
