% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrasts.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Compute estimated marginal means and contrasts from models.}
\usage{
get_contrasts(fit, ...)
}
\arguments{
\item{fit}{Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute estimated marginal means and contrasts from models.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
