% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get information about objects.}
\usage{
get_info(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Get information about objects. See the documentation for your object's class:
\itemize{
 \item{\link[=analyze.lmerModLmerTest]{analyze.merModLmerTest}}
 \item{\link[=analyze.glmerMod]{analyze.glmerMod}}
 \item{\link[=analyze.lm]{analyze.lm}}
 \item{\link[=analyze.lm]{analyze.stanreg}}
 }
 \itemize{
 \item{\link[=analyze.htest]{analyze.htest}}
 \item{\link[=analyze.aov]{analyze.aov}}
 }
\itemize{
 \item{\link[=analyze.fa]{analyze.fa}}
 \item{\link[=analyze.fa]{analyze.lavaan}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
