% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predicted.R
\name{get_predicted}
\alias{get_predicted}
\title{Compute predicted values from models.}
\usage{
get_predicted(fit, ...)
}
\arguments{
\item{fit}{Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute predicted values from models.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
