% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predicted.stanreg.R
\name{get_predicted.stanreg}
\alias{get_predicted.stanreg}
\title{Compute predicted values of stanreg models.}
\usage{
\method{get_predicted}{stanreg}(fit, refgrid = NULL,
  posterior_predict = FALSE, prob = 0.9, ...)
}
\arguments{
\item{fit}{A stanreg model.}

\item{refgrid}{Do the prediction on the actual dataframe or on a provided refgrid (created with emmeans). See the emmeans::ref_grid function.}

\item{posterior_predict}{Should the prediction be based on the posterior draws or not.}

\item{prob}{Probability of credible intervals (0.9 (default) will compute
5-95\% CI).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
dataframe with predicted values.
}
\description{
Compute predicted from a stanreg model.
}
\examples{
\dontrun{
library(psycho)
require(rstanarm)

fit <- rstanarm::stan_glm(Tolerating ~ Adjusting, data=affective, iter=500)

refgrid <- emmeans::ref_grid(fit, at=list(
    Adjusting=seq(min(affective$Adjusting), max(affective$Adjusting), length.out=2)))

predicted <- get_predicted(fit, refgrid)


fit <- rstanarm::stan_glm(Sex ~ Adjusting, data=affective, family="binomial", iter=500)

refgrid <- emmeans::ref_grid(fit, at=list(
    Adjusting=seq(min(affective$Adjusting), max(affective$Adjusting), length.out=5)))

predicted <- get_predicted(fit, refgrid, posterior_predict=FALSE)

}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
