% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_bf.R
\name{interpret_bf}
\alias{interpret_bf}
\title{Return the interpretation of a Bayes Factor.}
\usage{
interpret_bf(x, label_only = FALSE)
}
\arguments{
\item{x}{Bayes Factor.}

\item{label_only}{Keep only the size classification.}
}
\value{
The interpretation according to Jeffreys (1961).
}
\description{
Return the interpretation of a Bayes Factor.
}
\examples{
library(psycho)
interpret_bf(10)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
