% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_d.R
\name{interpret_d}
\alias{interpret_d}
\title{Return the interpretation of a Cohen's d or a standardized coefficient following Cohen (1988).}
\usage{
interpret_d(x)
}
\arguments{
\item{x}{Cohen's d value of standardized coefficient.}
}
\value{
The interpretation according to Cohen (1988).
}
\description{
Return the interpretation of a Cohen's d or a standardized coefficient following Cohen (1988).
}
\examples{
library(psycho)
interpret_d(-0.42)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
