% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_r.R
\name{interpret_r}
\alias{interpret_r}
\title{Correlation coefficient r interpreation.}
\usage{
interpret_r(x, direction = TRUE, strength = TRUE,
  rules = "cohen1988")
}
\arguments{
\item{x}{Correlation coefficient.}

\item{direction}{Return direction.}

\item{strength}{Return strength.}

\item{rules}{Can be "cohen1988" (default), "evans1996", or a custom list.}
}
\description{
Interpret r with a set of rules.
}
\examples{
library(psycho)
interpret_r(-0.42)

}
\seealso{
Page 88 of APA's 6th Edition
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
