% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_r.R
\name{interpret_r_posterior}
\alias{interpret_r_posterior}
\title{Correlation coefficient r interpreation for a posterior distribution.}
\usage{
interpret_r_posterior(posterior, rules = "cohen1988")
}
\arguments{
\item{posterior}{Posterior distribution of correlation coefficient.}

\item{rules}{Can be "cohen1988" (default) or "evans1996", or a custom list.}
}
\description{
Interpret r with a set of rules.
}
\examples{
library(psycho)
posterior <- rnorm(1000, 0.5, 0.5)
interpret_r_posterior(posterior)
}
\seealso{
Page 88 of APA's 6th Edition
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
