% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize (scale and reduce) numeric variables.}
\usage{
normalize(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Dataframe.
}
\description{
Select numeric variables and normalize (Z-score) them.
}
\examples{
df <- data.frame(Participant = as.factor(rep(1:50,each=2)), 
  Condition = base::rep_len(c("A", "B"), 100), V1 = rnorm(100, 30, .2), 
  V2 = runif(100, 3, 5))

dfZ <- normalize(df)

}
\author{
Dominique Makowski, \url{https://dominiquemakowski.github.io/}
}
