% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{odds_to_probs}
\alias{odds_to_probs}
\title{Convert (log)odds to probabilies.}
\usage{
odds_to_probs(odds, subset = NULL, except = NULL, log = TRUE)
}
\arguments{
\item{odds}{Odds values in vector or dataframe.}

\item{subset}{Character or list of characters of column names to be
transformed.}

\item{except}{Character or list of characters of column names to be excluded
from transformation.}

\item{log}{Are these Log odds (such as in logistic models)?}
}
\description{
Convert (log)odds to probabilies.
}
\examples{
library(psycho)
odds_to_probs(-1.45)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
