% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{print.psychobject}
\alias{print.psychobject}
\title{Print the results.}
\usage{
\method{print}{psychobject}(x, ...)
}
\arguments{
\item{x}{A psychobject class object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print the results.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
