% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_data_regression}
\alias{simulate_data_regression}
\title{Simulates data for single or multiple regression.}
\usage{
simulate_data_regression(sample = 10, coefs = 0, error = 0)
}
\arguments{
\item{sample}{Desired sample size.}

\item{coefs}{Desired theorethical coefs. Can be a single value or a list.}

\item{error}{The error (standard deviation of gaussian noise).}
}
\description{
Simulates data for single or multiple regression.
}
\details{
See https://stats.stackexchange.com/questions/59062/multiple-linear-regression-simulation
}
\examples{
library(psycho)

data <- simulate_data_regression(sample=50, coefs=c(0, 0.5), error=0.1)
lm(y ~ ., data=data)

}
\author{
TPArrow
}
