% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.numeric}
\alias{standardize.numeric}
\title{Standardize (scale and reduce) numeric variables.}
\usage{
\method{standardize}{numeric}(x, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Standardize (Z-score, "normalize") a vector.
}
\examples{
dfZ <- standardize(x=c(1, 4, 6, 2))


}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
