% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize (scale and reduce) numeric variables.}
\usage{
standardize(df, subset = NULL, except = NULL)
}
\arguments{
\item{df}{Dataframe.}

\item{subset}{Character or list of characters of column names to be
standardized.}

\item{except}{Character or list of characters of column names to be excluded
from standardized.}
}
\value{
Dataframe.
}
\description{
Select numeric variables and standardize (Z-score, "normalize") them.
}
\examples{
df <- data.frame(
  Participant = as.factor(rep(1:50,each=2)),
  Condition = base::rep_len(c("A", "B"), 100),
  V1 = rnorm(100, 30, .2),
  V2 = runif(100, 3, 5),
  V3 = rnorm(100, 100, 10)
  )

dfZ <- standardize(df)
dfZ <- standardize(df, except="V3")
dfZ <- standardize(df, except=c("V1", "V2"))
dfZ <- standardize(df, subset="V3")
dfZ <- standardize(df, subset=c("V1", "V2"))


}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
